WIP: implementing commands

+++ testSensor/datahandler.c
@@ -30,7 +30,7 @@ void data_ack_handler(uip_ipaddr_t *remote_addr, int remote_port, char *data, in
         return;
     }
 
-    printf("Data ack seq: %d\r\n", ack->ack_seq);
+    printf("Data ack seq: %u\r\n", (unsigned int) ack->ack_seq);
     data_seq_acked = ack->ack_seq;
     process_post(&test_data, PROCESS_EVENT_MSG, data);
 }
